+++ docs/changelog.rst
@@ -13,6 +13,11 @@ and this project adheres to `Semantic Versioning`_.
 -------------
 Added
 +++++
+- Added the "matches" operator (``<>REGEX``) which matches a version against
+  a regex (DOCS NEEDED)
+- Added the "in-range" operator (``=>V``) which matches a version against
+  a certain range of indexes (DOCS NEEDED)
+
 - Added the ability to specify ``--present-package`` multiple times using the
   same package name, but different versions. This is useful for when the
   ``:conflict-strat`` is set to ``inclusive``.